+++ src/activities/reader/EpubReaderActivity.h
@@ -51,6 +51,7 @@ class EpubReaderActivity final : public Activity {
     uint8_t embeddedStyle = 1;
     uint8_t hyphenationEnabled = 0;
     uint8_t textAntiAliasing = 1;
+    uint8_t textAntiAliasing1Bit = 0;
     uint8_t imageRendering = 0;
     uint8_t extraParagraphSpacing = 1;
     uint8_t forceParagraphIndents = 0;